py/demo-django/pyproject.toml
[project]
name = "demo-foo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.0",
    "niquests>=3.14.1",
]
